% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_race_data.R
\name{bl_download_2010_vars}
\alias{bl_download_2010_vars}
\title{Download 2010 Census variables at the block level}
\usage{
bl_download_2010_vars(state)
}
\arguments{
\item{state}{the two-letter abbreviation of the state to get data for.}
}
\value{
A data frame with the population counts
}
\description{
Downloads population and voting-age population by race at the block level
for the 2010 decennial Census. API key required:
\url{https://api.census.gov/data/key_signup.html}.
}
\examples{
\dontrun{
# API key required
bl_download_2010_vars("WA")
}

}
