% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_race_data.R
\name{bl_download_acs_vars}
\alias{bl_download_acs_vars}
\title{Download 2019 5-year ACS variables at the block group level}
\usage{
bl_download_acs_vars(state)
}
\arguments{
\item{state}{the two-letter abbreviation of the state to get data for.}
}
\value{
A data frame with the population counts
}
\description{
Downloads population and voting-age population by race at the block group
level for the 2019 5-year American Communities Survey.
}
\examples{
\donttest{
bl_download_acs_vars("WA")
}

}
