% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_2020.R
\name{bl_est_2020}
\alias{bl_est_2020}
\title{Estimate 2020 population from 2010-2019 populations}
\usage{
bl_est_2020(data, forecast_start = 2010)
}
\arguments{
\item{data}{the output of \code{\link[=bl_load_state]{bl_load_state()}}}

\item{forecast_start}{the year to use as the base year. E.g. if 2013, then
the average growth rate from 2013--2019 will be used.}
}
\value{
A modified data frame with just 2010, base year, and 2020 estimates.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Uses the average growth rate over the specified window to impute the 2020
population. Final estimates are rescaled to match the Census April 1, 2020
apportioment populations at the state level.
}
\examples{
\dontrun{
d = bl_load_state("WA")
bl_est_2020(d, 2010)
}

}
