% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_vars.R
\name{bl_harmonize_vars}
\alias{bl_harmonize_vars}
\title{Harmonize 2010 Decennial Census, 2019 ACS, and 2020 Block Population Estimates}
\usage{
bl_harmonize_vars(block_d, census_d, acs_d)
}
\arguments{
\item{block_d}{the output of \code{\link[=bl_est_2020]{bl_est_2020()}}}

\item{census_d}{the output of \code{\link[=bl_download_2010_vars]{bl_download_2010_vars()}} or similar}

\item{acs_d}{the output of \code{\link[=bl_download_acs_vars]{bl_download_acs_vars()}} or similar}
}
\value{
A data frame of 2020 block-level estimates for the variables in
\code{census_d}.
}
\description{
Creates estimates for race/ethnicity variables for blocks in 2020, based on
the 2010 Census (\code{\link[=bl_download_2010_vars]{bl_download_2010_vars()}}) and 2019 ACS
(\code{\link[=bl_download_acs_vars]{bl_download_acs_vars()}}) data.
}
\details{
The racial breakdown of total and voting-age population are proportionally
adjusted within block groups to match the ACS block-group percentages. Then
these percentages are applied to the newly estimated block populations.
Finally, the block race populations are rescaled so that the block
populations match the 2020 estimates. This is essentially a 1-round iterative
proportional fitting (IPF) procedure.
}
