% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_state.R
\name{bl_load_state}
\alias{bl_load_state}
\title{Download and extract population estimates for a state}
\usage{
bl_load_state(state, path = "https://www.fcc.gov/file/19314/download")
}
\arguments{
\item{state}{the two-letter abbreviation of the state to get data for.}

\item{path}{the path to the FCC data. Defaults to the online ZIP file, but
can point to a local zip file or an extracted CSV (from e.g. \code{\link[=bl_download_fcc]{bl_download_fcc()}}).}
}
\value{
A data frame with the population estimates.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\donttest{
bl_load_state("WA")
}

}
