% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockpop-package.R
\docType{package}
\name{blockpop-package}
\alias{blockpop}
\alias{blockpop-package}
\title{blockpop: Estimate Census Block Populations for 2020}
\description{
Uses FCC block-level population estimates from 2010--2019, which
    are based on new roads and map data, to estimate 2020 block populations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://corymccartan.github.io/blockpop}
}

}
\author{
\strong{Maintainer}: Cory McCartan \email{cmccartan@g.harvard.edu} (\href{https://orcid.org/0000-0002-6251-669X}{ORCID})

}
\keyword{internal}
